folder structure and new font added
reset styles added
