+++ src/fonts/font.css
@@ -1,0 +1,2 @@
+/* the Urbanist font used from google fonts */
+@import url("https://fonts.googleapis.com/css2?family=Urbanist:wght@400;500;600;700&display=swap");
