+++ src/global.css
@@ -1,3 +1,17 @@
+@import url(./fonts/font.css);
+
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+
+@layer base {
+  * {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+  }
+  body {
+    @apply font-urbanist;
+  }
+}
